Expression is an instance of Ord once again.
The Ordering of Expressions has been changed.  Constants come first
instead of last.  The unit tests now assume this is the case.

Displaying is broken because of this change, but displaying needs
to be rewritten anyway.

Unit tests of displaying now specify a character string as input,
not an explicit expression.  This gives end to end testing of
expression evaluation and parsing.

Some formatting fixups in the unit tests.

23 unit tests fail.