Found and fixed a very subtle bug in combining and eliminating terms of a sum.  Problem is that if the last two terms of a sum are combined, and generate a zero, then the zero ends up at the end of the sum. This zero was NOT REMOVED by subsequent recursions, because there was no pattern match that detected a zero as the last element of a sum.
This sneaked by because the "integration tests" of parsing and printing
somehow are passing the expression through the constructor functions
twice (I think).  Doign so corrects the bad sum on the second pass.

Problems that this raises:
(1) Integration tests are NOT sufficient; need tests of evaluation
(2) Same bug may exist in products as well
(3) Should only construct an expression once when it's parsed; where
    is the duplicate work happening and can it be removed?  It's
    important to do so because it could uncover more bugs that are
    currently hidden by double construction.